pyproject.toml
[project]
name = "BarcodeForge"
version = "1.0.0"
description = "A Python package template."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "dendropy>=4.5.2",
    "ete4>=4.3.0",
    "pathlib>=1.0.1",
    "rich-click>=1.6.0",
    "rich>=10.0.0",
    "six>=1.16.0",
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "biopython>=1.78",
    "altair>=5.0.0",
    "nextstrain-augur>=31.1.0",
]

[project.optional-dependencies]
dev-requirements = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "setuptools",
    "toml", # Added to parse pyproject.toml in tests
    "black", # Added for code formatting
]
test-requirements = [
    "pytest",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
barcodeforge = "barcodeforge.cli:cli"

[tool.barcodeforge.conda_dependencies]
# These packages are required and need to be installed separately using Conda.
# For example: conda install -c bioconda usher ucsc-fatovcf
packages = [
    "usher",  # (from bioconda channel)
    "ucsc-fatovcf",  # (from bioconda channel)
]
